el_per_state should be an integer, since it only takes on the values 1 or 2. Otherwise, we are passing up opportunities to use more accurate integer arithmetic.

+++ src/main/casida.F90
@@ -84,7 +84,7 @@ module casida_m
     integer           :: nst            !< total number of states
     integer           :: nik
     integer           :: sb_dim         !< number of spatial dimensions
-    FLOAT             :: el_per_state
+    integer           :: el_per_state
     character(len=80) :: trandens
     logical           :: triplet        !< use triplet kernel?
     logical           :: calc_forces    !< calculate excited-state forces
